feat: clean up views component form business logic + fix bug when copy of view did not uncrease aaset count
